Correct build issues related to issue #197
Problem with f16
Implicit conversions from float to int

+++ Source/FastMathFunctions/arm_atan2_f16.c
@@ -135,7 +135,7 @@ ARM_DSP_ATTRIBUTE arm_status arm_atan2_f16(float16_t y,float16_t x,float16_t *re
         }
         else
         {
-            if (signbit(y))
+            if (signbit((float)y))
             {
                *result= -(_Float16)PIF16;
             }
